+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - DynamicBrcode resource
 - PixKeyHolmes resource
 - PixKeyHolmes.Log sub-resource
+- PixInternalTransactionReport resource
+- PixInternalTransactionReport.Log sub-resource
 - PixPullRequest resource
 - PixPullRequest.Log sub-resource
 - Ledger resource